iOS: apply Aziz file/test/helper policy to the geometry-pin code
- Split MobileTerminalGeometryPin.swift into one type per file
  (MobileTerminalGridPin, MobileTerminalOutputChunk) and turn the no-case
  decision enum into a file-scope free function
  (mobileTerminalNextGridPin) with DocC docs on every public symbol.
- Convert MobileTerminalGeometryPinTests to Swift Testing.
- Move replayResponseGridPin off the MobileShellComposite god object to a
  file-scope private helper.

No behavior change.
